Revert "Show project list on project overview page"
