+++ mariadb-connector/src/main/java/life/qbic/finance/persistence/OffsetBasedRequest.java
@@ -1,4 +1,4 @@
-package life.qbic;
+package life.qbic.finance.persistence;
 
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
