+++ vaadinfrontend/src/main/java/life/qbic/datamanager/views/projectOverview/components/OfferSearchDialog.java
@@ -1,4 +1,4 @@
-package life.qbic.datamanager.views.project.overview.components;
+package life.qbic.datamanager.views.projectOverview.components;
 
 import static org.slf4j.LoggerFactory.getLogger;
